requirements.txt
pyaudio
openai
simpleaudio
RPi.GPIO

